cloud_run/requirements.txt
fastapi
uvicorn[standard]
google-cloud-storage
google-cloud-aiplatform
